+++ frontend2/lib/main.dart
@@ -13,10 +13,10 @@ import 'package:frontend2/screens/initial_setup_screen.dart';
 import 'package:frontend2/screens/main_navigation_screen.dart';
 import 'package:frontend2/screens/login_screen.dart';
 import 'package:frontend2/screens/mess_screen.dart';
-import 'package:frontend2/utilities/notifications.dart';
+import 'package:frontend2/providers/notification_provider.dart';
 import 'package:frontend2/constants/themes.dart';
-import 'package:frontend2/utilities/startupitem.dart';
-import 'package:frontend2/utilities/version_checker.dart';
+import 'package:frontend2/providers/mess_info_provider.dart';
+import 'package:frontend2/utils/version_checker.dart';
 import 'package:frontend2/services/festival_mode_service.dart';
 import 'package:hive_flutter/hive_flutter.dart';
 import 'package:provider/provider.dart';
@@ -116,6 +116,7 @@ Future<void> main() async {
         ChangeNotifierProvider(create: (_) => MessInfoProvider()),
         ChangeNotifierProvider(create: (_) => FeedbackProvider()),
         ChangeNotifierProvider(create: (_) => RoomCleaningProvider()),
+        ChangeNotifierProvider.value(value: globalNotificationProvider),
       ],
       child: MyApp(isLoggedIn: isLoggedIn, updateRequired: updateRequired),
     ),
@@ -147,8 +148,9 @@ class _MyAppState extends State<MyApp> with WidgetsBindingObserver {
   void initState() {
     super.initState();
     WidgetsBinding.instance.addObserver(this);
-    listenNotifications();
-    setNavigatorKey(navigatorKey); // Set global navigator key for notifications
+    context.read<NotificationProvider>().listenNotifications();
+    context.read<NotificationProvider>().setNavigatorKey(
+        navigatorKey); // Set global navigator key for notifications
     _connectivity = Connectivity();
 
     // Use `.map()` to transform the stream into a stream of ConnectivityResult
@@ -227,7 +229,7 @@ class _MyAppState extends State<MyApp> with WidgetsBindingObserver {
   void didChangeAppLifecycleState(AppLifecycleState state) {
     if (state == AppLifecycleState.resumed) {
       // The Filter Loop: Run on every foreground resume
-      filterAndLoadLocalAlerts();
+      context.read<NotificationProvider>().filterAndLoadLocalAlerts();
     }
   }
 }
